Update PotentialBuildsSetsResults to tsx with appropriate TS in new folder
